--- MeetU2/src/com/test/dto/EEventDTO.java
@@ -1,261 +1,0 @@
-package com.test.dto;
-
-public class EEventDTO
-{
-	private int event_id, evestatus_id, lmember_id, event_minattend
-	, event_maxattend, public_attendl, public_place, public_gender, public_age, public_eve, event_minm;
-
-	private String city_code, event_name, loc_geolat, loc_geolng, event_place, event_info, event_date, event_url
-	, group_name, public_id, public_name, category_code, category_content, lgroup_id;
-	
-	public String getLgroup_id()
-	{
-		return lgroup_id;
-	}
-
-	public void setLgroup_id(String lgroup_id)
-	{
-		this.lgroup_id = lgroup_id;
-	}
-
-	public String getCategory_code()
-	{
-		return category_code;
-	}
-
-	public void setCategory_code(String category_code)
-	{
-		this.category_code = category_code;
-	}
-
-	public String getCategory_content()
-	{
-		return category_content;
-	}
-
-	public void setCategory_content(String category_content)
-	{
-		this.category_content = category_content;
-	}
-
-	public String getPublic_id()
-	{
-		return public_id;
-	}
-
-	public void setPublic_id(String public_id)
-	{
-		this.public_id = public_id;
-	}
-
-	public String getPublic_name()
-	{
-		return public_name;
-	}
-
-	public void setPublic_name(String public_name)
-	{
-		this.public_name = public_name;
-	}
-
-	public String getGroup_name()
-	{
-		return group_name;
-	}
-
-	public void setGroup_name(String group_name)
-	{
-		this.group_name = group_name;
-	}
-
-	public int getEvent_id()
-	{
-		return event_id;
-	}
-
-	public void setEvent_id(int event_id)
-	{
-		this.event_id = event_id;
-	}
-
-	public int getEvestatus_id()
-	{
-		return evestatus_id;
-	}
-
-	public void setEvestatus_id(int evestatus_id)
-	{
-		this.evestatus_id = evestatus_id;
-	}
-
-	public int getLmember_id()
-	{
-		return lmember_id;
-	}
-
-	public void setLmember_id(int lmember_id)
-	{
-		this.lmember_id = lmember_id;
-	}
-
-	public int getEvent_minattend()
-	{
-		return event_minattend;
-	}
-
-	public void setEvent_minattend(int event_minattend)
-	{
-		this.event_minattend = event_minattend;
-	}
-
-	public int getEvent_maxattend()
-	{
-		return event_maxattend;
-	}
-
-	public void setEvent_maxattend(int event_maxattend)
-	{
-		this.event_maxattend = event_maxattend;
-	}
-
-	public int getPublic_attendl()
-	{
-		return public_attendl;
-	}
-
-	public void setPublic_attendl(int public_attendl)
-	{
-		this.public_attendl = public_attendl;
-	}
-
-	public int getPublic_place()
-	{
-		return public_place;
-	}
-
-	public void setPublic_place(int public_place)
-	{
-		this.public_place = public_place;
-	}
-
-	public int getPublic_gender()
-	{
-		return public_gender;
-	}
-
-	public void setPublic_gender(int public_gender)
-	{
-		this.public_gender = public_gender;
-	}
-
-	public int getPublic_age()
-	{
-		return public_age;
-	}
-
-	public void setPublic_age(int public_age)
-	{
-		this.public_age = public_age;
-	}
-
-	public int getPublic_eve()
-	{
-		return public_eve;
-	}
-
-	public void setPublic_eve(int public_eve)
-	{
-		this.public_eve = public_eve;
-	}
-
-	public int getEvent_minm()
-	{
-		return event_minm;
-	}
-
-	public void setEvent_minm(int event_minm)
-	{
-		this.event_minm = event_minm;
-	}
-
-	public String getEvent_url()
-	{
-		return event_url;
-	}
-
-	public void setEvent_url(String event_url)
-	{
-		this.event_url = event_url;
-	}
-
-	public String getCity_code()
-	{
-		return city_code;
-	}
-
-	public void setCity_code(String city_code)
-	{
-		this.city_code = city_code;
-	}
-
-	public String getEvent_name()
-	{
-		return event_name;
-	}
-
-	public void setEvent_name(String event_name)
-	{
-		this.event_name = event_name;
-	}
-
-	public String getLoc_geolat()
-	{
-		return loc_geolat;
-	}
-
-	public void setLoc_geolat(String loc_geolat)
-	{
-		this.loc_geolat = loc_geolat;
-	}
-
-	public String getLoc_geolng()
-	{
-		return loc_geolng;
-	}
-
-	public void setLoc_geolng(String loc_geolng)
-	{
-		this.loc_geolng = loc_geolng;
-	}
-
-	public String getEvent_place()
-	{
-		return event_place;
-	}
-
-	public void setEvent_place(String event_place)
-	{
-		this.event_place = event_place;
-	}
-
-	public String getEvent_info()
-	{
-		return event_info;
-	}
-
-	public void setEvent_info(String event_info)
-	{
-		this.event_info = event_info;
-	}
-
-	public String getEvent_date()
-	{
-		return event_date;
-	}
-
-	public void setEvent_date(String event_date)
-	{
-		this.event_date = event_date;
-	}
-
-}
